+++ CHANGELOG.md
@@ -21,5 +21,6 @@
 * **Integration-Library Salesforce:** Set to Create Chat from Salesforce. We need to create an endpoint to create and initialize a chat with Salesforce. Closes #CSF-155
 * **Integration-interconnection:** Interconnection for Message Handler. Add handler to send message to salesforce . Closes #CSF-82
 * **Integration-Manager:** Manage blocked user. If user is blocked from salesforce, we send it to the from-sf-blocked state. Closes #CSF-157
+* **Integration-Library Salesforce:** Set to request chat with Salesforce (Add Case) We add the caseId and the contactId to the chat request. Closes #CSF-158
 * **Integration-webhook:**  Set webhook in integrations. Add webhook endpoint  to save message form integrations Closes #CNF-140